requirement.txt
fastapi
uvicorn
beanie
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
email-validator
stripe

